+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [7.6.0] 2025-06-10
+
+## Added
+
+- [TD-7221] Highlight the headers of the template fields in Concepts download file
+- [TD-6468] Multi language upload and download business concepts
+
+## Fixed
+
+- [TD-7243] Indexing translatable fields without default value
+
 ## [7.5.2] 2025-06-05
 
 ### Added